Create a tests directory and move all unit tests there for organized file structure
