+++ src/project.js
@@ -17,27 +17,6 @@ const defaultProject = new Project('Default');
 
 projects.push(defaultProject);
 
-// const getProjectsFromLocalStorage = () => {
-//   let projects;
-//   const localStorageProjects = localStorage.getItem('projects');
-//   if (localStorageProjects === null) {
-//     projects = [];
-//     projects.push(defaultProject);
-//     localStorage.setItem('projects', JSON.stringify(projects));
-//     projects = JSON.parse(projects);
-//   } else {
-//     projects = JSON.parse(localStorageProjects);
-//   }
-//   return projects;
-// };
-
-// const addProjectToLocalStorage = (newProject) => {
-//   const projects = getProjectsFromLocalStorage();
-//   projects.push(newProject);
-
-//   localStorage.setItem('projects', JSON.stringify(projects));
-// };
-
 const allprojects = document.getElementById('all-projects');
 
 const createProjectCard = (project) => {
@@ -60,7 +39,6 @@ const createProjectCard = (project) => {
 };
 
 const displayAllProjects = () => {
-  // const projects = getProjectsFromLocalStorage();
   projects.forEach((project) => {
     const projectCard = createProjectCard(project);
     allprojects.appendChild(projectCard);
@@ -70,7 +48,6 @@ const displayAllProjects = () => {
 const createProject = (name) => {
   const newProject = new Project(name);
   projects.push(newProject);
-  // addProjectToLocalStorage(newProject);
   allprojects.textContent = '';
   displayAllProjects();
   return newProject;
@@ -82,4 +59,4 @@ const localStorageOnLoad = () => {
 
 document.addEventListener('DOMContentLoaded', localStorageOnLoad);
 
-export { Project, createProject };
+export { Project, createProject, projects };
